+++ LibNoise/Operator/Translate.cs
@@ -6,12 +6,12 @@ namespace LibNoise.Operator
 {
     public class Translate : ModuleBase
     {
-        private readonly float tx;
-        private readonly float ty;
-        private readonly float tz;
+        private readonly double tx;
+        private readonly double ty;
+        private readonly double tz;
         private readonly ModuleBase original;
 
-        public Translate(float tx, float ty, float tz, ModuleBase original)
+        public Translate(double tx, double ty, double tz, ModuleBase original)
         {
             this.tx = tx;
             this.ty = ty;
@@ -19,7 +19,7 @@ namespace LibNoise.Operator
             this.original = original;
         }
 
-        public override float GetValue(float x, float y, float z)
+        public override double GetValue(double x, double y, double z)
         {
             return original.GetValue(tx + x, ty + y, tz + z);
         }
